+++ src/main/java/spot/spot/domain/chat/chatparticipant/entity/ChatParticipant.java
@@ -14,6 +14,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import spot.spot.domain.chat.chatroom.entity.ChatRoom;
+import spot.spot.domain.member.entity.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
