+++ src/main/java/spot/spot/domain/chat/chatroom/entity/ChatRoom.java
@@ -15,6 +15,7 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import spot.spot.domain.job.entity.Job;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
