+++ src/main/java/spot/spot/domain/chat/message/entity/Message.java
@@ -16,6 +16,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import spot.spot.domain.chat.chatroom.entity.ChatRoom;
+import spot.spot.domain.member.entity.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
