+++ src/main/java/spot/spot/domain/chat/messagestatus/entity/MessageStatus.java
@@ -1,4 +1,4 @@
-package spot.spot.chat.messagestatus.domain;
+package spot.spot.domain.chat.messagestatus.entity;
 
 import org.hibernate.annotations.Comment;
 
@@ -15,6 +15,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import spot.spot.domain.chat.chatroom.entity.ChatRoom;
 import spot.spot.domain.chat.message.entity.Message;
+import spot.spot.domain.member.entity.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
